requirements.txt
attrs==20.3.0
Authlib==0.15.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
cycler==0.10.0
h11==0.12.0
httpcore==0.12.3
httpx==0.17.1
idna==2.10
iniconfig==1.1.1
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.4.1
numpy==1.20.2
packaging==20.9
panda==0.3.1
pandas==1.2.3
pandas-datareader==0.9.0
Pillow==8.2.0
pluggy==0.13.1
prompt-toolkit==3.0.18
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rfc3986==1.4.0
selenium==3.141.0
six==1.15.0
sniffio==1.2.0
tda-api==1.1.8
toml==0.10.2
urllib3==1.26.4
wcwidth==0.2.5
websockets==8.1
